indexer: narrow signer_id/gas to u64, *_index/vout to u32

+++ core/indexer-types/src/lib.rs
@@ -347,8 +347,7 @@ pub enum Event {
 pub struct Input {
     #[ts(as = "String")]
     pub previous_output: bitcoin::OutPoint,
-    #[ts(type = "number")]
-    pub input_index: i64,
+    pub input_index: u32,
     #[ts(as = "String")]
     pub x_only_pubkey: XOnlyPublicKey,
     pub insts: Insts,
@@ -359,8 +358,7 @@ pub struct Input {
 pub struct Transaction {
     #[ts(type = "string")]
     pub txid: Txid,
-    #[ts(type = "number")]
-    pub index: i64,
+    pub index: u32,
     pub inputs: Vec<Input>,
     /// OP_RETURN directives, one entry per reveal input that carries one.
     pub op_return_data: Vec<OpReturnEntry>,
@@ -382,7 +380,7 @@ pub struct Block {
 #[ts(export, export_to = "../../../sdk/src/bindings.d.ts")]
 pub struct OutPoint {
     pub txid: String,
-    pub vout: u64,
+    pub vout: u32,
 }
 
 #[derive(Debug, Clone, Hash, Serialize, Deserialize, TS)]
@@ -407,7 +405,7 @@ impl core::fmt::Display for OutPoint {
 pub struct ContractAddress {
     pub name: String,
     pub height: u64,
-    pub tx_index: u64,
+    pub tx_index: u32,
 }
 
 contract_address!(ContractAddress);
@@ -417,10 +415,8 @@ contract_address!(ContractAddress);
 pub struct OpMetadata {
     #[ts(as = "String")]
     pub previous_output: bitcoin::OutPoint,
-    #[ts(type = "number")]
-    pub input_index: i64,
-    #[ts(type = "number")]
-    pub op_index: i64,
+    pub input_index: u32,
+    pub op_index: u32,
     #[ts(type = "number")]
     pub signer_id: u64,
     pub payment: Payment,
@@ -612,8 +608,7 @@ pub struct TransactionRow {
     pub height: u64,
     #[ts(type = "number | null")]
     pub confirmed_height: Option<u64>,
-    #[ts(type = "number | null")]
-    pub tx_index: Option<i64>,
+    pub tx_index: Option<u32>,
     #[ts(type = "number | null")]
     pub batch_height: Option<u64>,
 }
@@ -626,12 +621,11 @@ pub struct ContractListRow {
     pub name: String,
     #[ts(type = "number")]
     pub height: u64,
-    #[ts(type = "number")]
-    pub tx_index: i64,
+    pub tx_index: u32,
     #[ts(type = "number")]
     pub size: i64,
     #[ts(type = "number | null")]
-    pub signer_id: Option<i64>,
+    pub signer_id: Option<u64>,
 }
 
 #[serde_as]
@@ -642,12 +636,10 @@ pub struct PaginationMeta {
     #[serde(skip_serializing_if = "Option::is_none")]
     #[serde_as(as = "Option<DisplayFromStr>")]
     pub next_cursor: Option<i64>,
-    #[ts(type = "number | null")]
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub next_offset: Option<i64>,
+    pub next_offset: Option<u32>,
     pub has_more: bool,
-    #[ts(type = "number")]
-    pub total_count: i64,
+    pub total_count: u32,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, TS)]
@@ -763,10 +755,8 @@ pub enum OpWithResult {
         error_message: Option<String>,
     },
     Rejected {
-        #[ts(type = "number")]
-        input_index: i64,
-        #[ts(type = "number")]
-        op_index: i64,
+        input_index: u32,
+        op_index: u32,
         error_message: Option<String>,
     },
 }
@@ -793,14 +783,14 @@ impl OpWithResult {
         }
     }
 
-    pub fn input_index(&self) -> i64 {
+    pub fn input_index(&self) -> u32 {
         match self {
             OpWithResult::Materialized { op, .. } => op.metadata.input_index,
             OpWithResult::Rejected { input_index, .. } => *input_index,
         }
     }
 
-    pub fn op_index(&self) -> i64 {
+    pub fn op_index(&self) -> u32 {
         match self {
             OpWithResult::Materialized { op, .. } => op.metadata.op_index,
             OpWithResult::Rejected { op_index, .. } => *op_index,
@@ -868,17 +858,13 @@ pub struct ResultRow {
     pub id: i64,
     #[ts(type = "number")]
     pub height: u64,
-    #[ts(type = "number | null")]
-    pub tx_index: Option<i64>,
-    #[ts(type = "number | null")]
-    pub input_index: Option<i64>,
-    #[ts(type = "number | null")]
-    pub op_index: Option<i64>,
-    #[ts(type = "number")]
-    pub result_index: i64,
+    pub tx_index: Option<u32>,
+    pub input_index: Option<u32>,
+    pub op_index: Option<u32>,
+    pub result_index: u32,
     pub func: String,
     #[ts(type = "number")]
-    pub gas: i64,
+    pub gas: u64,
     /// Outcome category for this op. `Ok` for successful calls (regardless
     /// of whether the contract returned `ok(...)` or just a value); the
     /// failure variants distinguish what went wrong.
@@ -887,12 +873,12 @@ pub struct ResultRow {
     pub contract: String,
     pub txid: Option<String>,
     #[ts(type = "number")]
-    pub signer_id: i64,
+    pub signer_id: u64,
     /// Who funded gas for this op. Equals `signer_id` for self-pay ops;
     /// for BLS-aggregate sponsored ops it's the publisher's signer_id.
     /// `null` for ops that don't go through user-side gas accounting.
     #[ts(type = "number | null")]
-    pub payer_signer_id: Option<i64>,
+    pub payer_signer_id: Option<u64>,
 }
 
 /// One OP_RETURN directive bound to the reveal input it applies to:
